backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-docx==1.1.2
